Move getEventsForMultipleProviders to helpers file
