+++ packages/financial-templates-lib/src/clients/InsuredBridgeL2Client.ts
@@ -4,7 +4,9 @@
 import { getAbi } from "@uma/contracts-node";
 import type { BridgeDepositBoxWeb3 } from "@uma/contracts-node";
 import Web3 from "web3";
-import { EventData } from "web3-eth-contract";
+import type { EventData } from "web3-eth-contract";
+import { EventSearchOptions, getEventsForMultipleProviders } from "../helpers/contractEvents";
+
 const { toChecksumAddress } = Web3.utils;
 import type { Logger } from "winston";
 
@@ -21,13 +23,6 @@ export interface Deposit {
   quoteTimestamp: number;
   depositContract: string;
 }
-
-type EventSearchOptions = {
-  fromBlock: number;
-  toBlock: number;
-  filter?: any; // Object, allows caller to filter events by indexed paramateres. e.g. {filter: {myNumber: [12,13]}}
-  // filters all events where "myNumber" is 12 or 13.
-};
 
 export class InsuredBridgeL2Client {
   public bridgeDepositBox: BridgeDepositBoxWeb3;
@@ -131,7 +126,7 @@ export class InsuredBridgeL2Client {
   }
 
   async getFundsDepositedEvents(eventSearchOptions: EventSearchOptions): Promise<EventData[]> {
-    const eventsData = await this.getEventsForMultipleProviders(
+    const eventsData = await getEventsForMultipleProviders(
       [this.l2Web3].concat(this.fallbackL2Web3s),
       this.bridgeDepositBox.options.jsonInterface,
       this.bridgeDepositAddress,
@@ -151,7 +146,7 @@ export class InsuredBridgeL2Client {
   }
 
   async getWhitelistTokenEvents(eventSearchOptions: EventSearchOptions): Promise<EventData[]> {
-    const eventsData = await this.getEventsForMultipleProviders(
+    const eventsData = await getEventsForMultipleProviders(
       [this.l2Web3].concat(this.fallbackL2Web3s),
       this.bridgeDepositBox.options.jsonInterface,
       this.bridgeDepositAddress,
@@ -170,79 +165,6 @@ export class InsuredBridgeL2Client {
     return eventsData.events;
   }
 
-  /**
-   * Fetches specified contract event data for all input web3 providers. Returns false if any of the events found with
-   * one provider are NOT matched exactly in ALL of the other providers' event arrays.
-   * @param web3s Web3 providers to check for target event.
-   * @param contractAbi Contract ABI to query target event on.
-   * @param contractAddress Contract address to query target event on.
-   * @param eventName Name of target event.
-   * @param eventSearchOptions Target event search options. See here for more details: https://web3js.readthedocs.io/en/v1.5.2/web3-eth-contract.html#getpastevents
-   * @returns Object containing success of event query, missing events not found in all providers, and all event data
-   */
-  async getEventsForMultipleProviders(
-    web3s: Web3[],
-    contractAbi: any[],
-    contractAddress: string,
-    eventName: string,
-    eventSearchOptions: EventSearchOptions
-  ): Promise<{ missingEvents: EventData[]; events: EventData[] }> {
-    const allProviderEvents = await Promise.all(
-      web3s.map((_web3) => {
-        const _contract = new _web3.eth.Contract(contractAbi, contractAddress);
-        return _contract.getPastEvents(eventName, eventSearchOptions);
-      })
-    );
-
-    // Associate each event with a key uniquely identifying the event. We'll use this key in the next step to determine
-    // which events were returned by all providers.
-    type UniqueEventData = { [uniqueEventKey: string]: EventData };
-    const uniqueEventsForProvider: UniqueEventData[] = [];
-    // [index of web3Provider => {eventKey => event}]
-    const uniqueEvents: UniqueEventData = {};
-    // Union map of ALL unique events across all providers: [{eventKey => event}]. We'll use this map as the starting
-    // point for all of the matched events that we'll return from this function.
-
-    const _getUniqueEventKey = (event: EventData): string => {
-      return JSON.stringify({
-        transactionHash: event.transactionHash,
-        transactionIndex: event.transactionIndex,
-        logIndex: event.logIndex,
-        returnValues: event.returnValues,
-        address: event.address,
-      });
-    };
-    allProviderEvents.forEach((eventDataForProvider, i) => {
-      uniqueEventsForProvider[i] = {};
-      eventDataForProvider.forEach((event) => {
-        const uniqueEventKey = _getUniqueEventKey(event);
-        // Add event for this provider.
-        uniqueEventsForProvider[i][uniqueEventKey] = event;
-        // Add event to union map if we haven't seen it before.
-        uniqueEvents[uniqueEventKey] = event;
-      });
-    });
-
-    // Store only the events returned by ALL providers.
-    const eventKeysReturnedByAllProviders: string[] = [];
-    const missingEventKeys: string[] = [];
-    Object.keys(uniqueEvents).forEach((eventKey: string) => {
-      let eventFoundInAllProviders = true;
-      for (let providerIndex = 1; providerIndex < uniqueEventsForProvider.length; providerIndex++) {
-        if (uniqueEventsForProvider[providerIndex][eventKey] === undefined) {
-          eventFoundInAllProviders = false;
-          break;
-        }
-      }
-      if (eventFoundInAllProviders) eventKeysReturnedByAllProviders.push(eventKey);
-      else missingEventKeys.push(eventKey);
-    });
-
-    return {
-      missingEvents: missingEventKeys.map((eventKey) => uniqueEventsForProvider[0][eventKey]),
-      events: eventKeysReturnedByAllProviders.map((eventKey) => uniqueEventsForProvider[0][eventKey]),
-    };
-  }
   generateDepositHash = (depositData: Deposit): string => {
     const depositDataAbiEncoded = this.l2Web3.eth.abi.encodeParameters(
       ["uint256", "uint64", "address", "address", "uint256", "uint64", "uint64", "uint32", "address"],
